feat(JsonSchema)!: changed out decimal for double to allow for bigger numbers
BREAKING CHANGE: this changes the type of 3 properties of JsonSchema from `decimal` to `double`

+++ src/LEGO.AsyncAPI/Models/AsyncApiSchema.cs
@@ -37,7 +37,7 @@ namespace LEGO.AsyncAPI.Models
         /// <summary>
         /// follow JSON Schema definition: https://json-schema.org/draft-07/json-schema-release-notes.html.
         /// </summary>
-        public decimal? Maximum { get; set; }
+        public double? Maximum { get; set; }
 
         /// <summary>
         /// follow JSON Schema definition: https://json-schema.org/draft-07/json-schema-release-notes.html.
@@ -47,7 +47,7 @@ namespace LEGO.AsyncAPI.Models
         /// <summary>
         /// follow JSON Schema definition: https://json-schema.org/draft-07/json-schema-release-notes.html.
         /// </summary>
-        public decimal? Minimum { get; set; }
+        public double? Minimum { get; set; }
 
         /// <summary>
         /// follow JSON Schema definition: https://json-schema.org/draft-07/json-schema-release-notes.html.
@@ -73,7 +73,7 @@ namespace LEGO.AsyncAPI.Models
         /// <summary>
         /// follow JSON Schema definition: https://json-schema.org/draft-07/json-schema-release-notes.html.
         /// </summary>
-        public decimal? MultipleOf { get; set; }
+        public double? MultipleOf { get; set; }
 
         /// <summary>
         /// Follow JSON Schema definition: https://tools.ietf.org/html/draft-fge-json-schema-validation-00
